requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==1.3.0
python-multipart==0.0.6